fix: Don't override help strings
strcutopt lets you define the help string via doc-comments.  Recently,
they cleaned this up which introduced a bug where a chained struct wins
out over the parent.

So moving our doc-comment to the lib to workaround it.  We should have
the majority of the documentation on the lib anyways.

Fixed #20