fix!: Give control over help text
A runtime default means we can't change our help text off of it.  We
could accept the default as a const generics and match off of that for
help text but the user might want more control and, for now, we'd have
to expose all of that as integers.  So we're offering a trait with
zero-sized types to control the default level and different clap
settings, like help text.