Add optional support for `tracing` besides `log`
As an optional feature, add support for getting the verbosity level as
a `tracing::Level` to pass to Tokio's `tracing` library, very similar
to the existing support for getting the verbosity level as a
`log::Level`.

A `tracing` equivalent of the existing `Verbosity::log_level_filter`
method easily could be added as well:

```rust
/// Get the log level filter, as defined by the `tracing` crate.
pub fn tracing_level_filter(&self) -> tracing::level_filters::LevelFilter {
    self.tracing_level().into()
}
```

However, *that* method's body is so simple that I question whether
adding it would be worth your potentially needing to update it if
`tracing`, for example, moves the `LevelFilter` type out of that
`level_filters` module.